Move tests into their own files
